requirements.txt
arabic-reshaper==2.1.4
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.12.13
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.6
Django==4.1.5
django-import-export==3.0.2
django-oauth-toolkit==2.2.0
django-pandas==0.6.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
idna==3.4
isort==5.11.4
jwcrypto==1.4.2
lazy-object-proxy==1.8.0
lxml==4.9.2
MarkupPy==1.14
mccabe==0.7.0
numpy==1.24.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
pandas==1.5.2
Pillow==9.4.0
platformdirs==2.6.2
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
pylint==2.15.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tablib==3.3.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
